@@ -4,19 +4,19 @@
   "description": "Demo for billboard reflection.",
   "main": "src/index.js",
   "dependencies": {
-    "@react-three/drei": "^7.3.1",
-    "@react-three/fiber": "6.2.2",
-    "@react-three/postprocessing": "^2.0.5",
+    "@react-three/drei": "^9.4.2",
+    "@react-three/fiber": "^8.0.11",
+    "@react-three/postprocessing": "^2.3.2",
     "@rollup/plugin-babel": "^5.3.0",
     "esbuild": "^0.12.19",
     "leva": "^0.9.13",
-    "postprocessing": "^6.22.3",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-scripts": "4.0.3",
+    "postprocessing": "^6.26.3",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-scripts": "^5.0.1",
     "rollup-plugin-cleanup": "^3.2.1",
     "rollup-plugin-terser": "^7.0.2",
-    "three": "0.129.0"
+    "three": "^0.139.2"
   },
   "devDependencies": {
     "@ampproject/rollup-plugin-closure-compiler": "^0.27.0",
