@@ -1,6 +1,11 @@
-import ReactDOM from 'react-dom'
+import { createRoot } from 'react-dom/client'
 import React from 'react'
 import './demo/styles.css'
 import App from './demo/App'
 
-ReactDOM.render(<App />, document.getElementById('root'))
+const rootElement = document.getElementById('root');
+const root = createRoot(rootElement);
+
+root.render(
+    <App/>
+  );
